修改bug

+++ msit/components/llm/msit_llm/opcheck/check_case/self_attention.py
@@ -263,7 +263,7 @@ class OpcheckUnpadSelfAttentionOperation(operation_test.OperationTest):
                     no_compress_mask = torch.triu(no_compress_mask, 1)
                     no_compress_mask *= -10000.0
                     if is_mask:
-                        self.in_tensors[3] = no_compress_mask
+                        self.in_tensors[3] = no_compress_mask.to(self.in_tensors[3].dtype)
                     score = score + no_compress_mask[:, :q_s, :kv_s]
                 else:
                     score = score + mask[:, :q_s, :kv_s] * post_mask_coff
